+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Fixed bug causing user configuration to be only active for the first
+  `Readline` context created
 - Made sure to release initially `libreadline` allocated buffers to
   prevent one-time memory leak